+++ opensampl/vendors/adva.py
@@ -16,9 +16,9 @@ from opensampl.metrics import METRICS
 from opensampl.references import REF_TYPES
 from opensampl.vendors.base_probe import BaseProbe
 from opensampl.vendors.constants import VENDORS, ProbeKey
+from opensampl.mixins.random_data import RandomDataMixin
 
-
-class AdvaProbe(BaseProbe):
+class AdvaProbe(BaseProbe, RandomDataMixin):
     """ADVA Probe Object"""
 
     timestamp: datetime
@@ -32,7 +32,7 @@ class AdvaProbe(BaseProbe):
         r"(?P<hour>\d+)-(?P<minute>\d+)-(?P<second>\d+)\.txt(?:\.gz)?"
     )
 
-    class RandomDataConfig(BaseProbe.RandomDataConfig):
+    class RandomDataConfig(RandomDataMixin.RandomDataConfig):
         """Model for storing random data generation configurations as provided by CLI or YAML"""
 
         # Time series parameters
