+++ opensampl/vendors/microchip/twst.py
@@ -17,19 +17,20 @@ from pydantic import Field
 from sqlalchemy.exc import IntegrityError
 
 from opensampl.load_data import load_probe_metadata
+from opensampl.mixins.random_data import RandomDataMixin
 from opensampl.metrics import METRICS
 from opensampl.references import REF_TYPES
 from opensampl.vendors.base_probe import BaseProbe
 from opensampl.vendors.constants import VENDORS, ProbeKey
 
 
-class MicrochipTWSTProbe(BaseProbe):
+class MicrochipTWSTProbe(BaseProbe, RandomDataMixin):
     """MicrochipTWST Probe Object"""
 
     vendor = VENDORS.MICROCHIP_TWST
     MEASUREMENTS: ClassVar = {"meas:offset": METRICS.PHASE_OFFSET, "tracking:ebno": METRICS.EB_NO}
 
-    class RandomDataConfig(BaseProbe.RandomDataConfig):
+    class RandomDataConfig(RandomDataMixin.RandomDataConfig):
         """Model for storing random data generation configurations as provided by CLI or YAML"""
 
         # Time series parameters
